+++ spec/cobrato/entities/charge_config_spec.rb
@@ -14,7 +14,7 @@ describe Cobrato::Entities::ChargeConfig do
       end_number: 1000,
       status: "pending",
       registered_charges: true,
-      agreement_number: 4576361,
+      remittance_agreement_code: 4576361,
       initial_remittance_number: 1,
       remittance_cnab_pattern: 240,
       transmission_code: "1234567"
@@ -28,7 +28,7 @@ describe Cobrato::Entities::ChargeConfig do
     :id, :type, :name, :status,
     # billet specifics
     :bank_account_id, :portfolio_code, :agreement_code, :agreement_code_digit, :initial_number, :next_number,
-    :end_number, :registered_charges, :agreement_number, :initial_remittance_number, :current_remittance_number,
+    :end_number, :registered_charges, :remittance_agreement_code, :initial_remittance_number, :current_remittance_number,
     :remittance_cnab_pattern, :transmission_code,
     # payment gateway specifics
     :gateway_id, :gateway_key, :gateway_name, :use_avs]